fix information loss in property test
The test did first encode a message to binary and them check whether
and dec+enc on that binary returned the same binary.
If the initial encode already lost infromation (e.g. due to too small
data type) the dec+enc check would not catch that.
Instead compare the result of an enc+dec test to the original message.